2674: read bulk allocation update files on a background job
